ml-microservices-platform/services/image-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
boto3==1.28.62
Pillow==10.0.0
prometheus-client==0.17.1